+++ src/pages/admin/AdminContactsPage.tsx
@@ -4,139 +4,22 @@ import {
   ListItem,
   ListItemAvatar,
   ListItemText,
-  NativeSelect,
   IconButton,
   Typography,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogContentText,
-  DialogActions,
-  Button,
-  TextField,
-  Snackbar,
+   
 } from '@mui/material';
-import emailjs from '@emailjs/browser';
-import { CustomWindow } from '../../api/api';
 import { Avatar, CircularProgress } from '@mui/joy';
 import React from 'react';
 
 import { useAdminQueries } from '../../hooks';
 import { Contact } from '../../types/contact.types';
 
-import { useState } from 'react';
 import { WhatsApp, Email, Phone } from '@mui/icons-material';
 import { CustomAlert } from '../../utils/customAlert';
 
-interface ContactUs {
-  name: string;
-  email: string;
-  message: string;
-}
-
-interface SnackbarAlert {
-  open: boolean;
-  message: string;
-  severity: 'success' | 'error' | 'warning' | 'info';
-}
-
 const AdminContactsPage: React.FC = () => {
-  const [openContactDialog, setOpenContactDialog] = useState<boolean>(false);
-  const [isSubmitting, setIsSubmitting] = useState<boolean>(false);
-  const [contactForm, setContactForm] = useState<ContactUs>({
-    name: '',
-    email: '',
-    message: '',
-  });
+  
 
-  const [snackbar, setSnackbar] = useState<SnackbarAlert>({
-    open: false,
-    message: '',
-    severity: 'success',
-  });
-
-  const handleContactDialogOpen = () => {
-    setOpenContactDialog(true);
-  };
-
-  const handleContactDialogClose = () => {
-    setOpenContactDialog(false);
-    setContactForm({
-      name: '',
-      email: '',
-      message: '',
-    });
-  };
-
-  const handleContactFormChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setContactForm({
-      ...contactForm,
-      [event.target.name]: event.target.value,
-    });
-  };
-
-  const handleContactSubmit = async () => {
-    // Basic validation
-    if (!contactForm.name.trim() || !contactForm.email.trim() || !contactForm.message.trim()) {
-      setSnackbar({
-        open: true,
-        message: 'Please fill in all fields',
-        severity: 'error',
-      });
-      return;
-    }
-
-    // Basic email validation
-    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    if (!emailRegex.test(contactForm.email)) {
-      setSnackbar({
-        open: true,
-        message: 'Please enter a valid email address',
-        severity: 'error',
-      });
-      return;
-    }
-
-    setIsSubmitting(true);
-
-    try {
-      const templateParams = {
-        from_name: contactForm.name,
-        from_email: contactForm.email,
-        message: contactForm.message,
-        to_email: 'taxibee2024@gmail.com',
-      };
-
-      const VITE_EMAILJS_PUBLIC_KEY = (window as CustomWindow).__ENV__?.VITE_EMAILJS_PUBLIC_KEY;
-      const VITE_EMAILJS_SERVICE_ID = (window as CustomWindow).__ENV__?.VITE_EMAILJS_SERVICE_ID;
-      const VITE_EMAILJS_TEMPLATE_ID = (window as CustomWindow).__ENV__?.VITE_EMAILJS_TEMPLATE_ID;
-
-      await emailjs.send(
-        VITE_EMAILJS_SERVICE_ID!,
-        VITE_EMAILJS_TEMPLATE_ID!,
-        templateParams,
-        VITE_EMAILJS_PUBLIC_KEY!
-      );
-
-      setSnackbar({
-        open: true,
-        message: 'Message sent successfully!',
-        severity: 'success',
-      });
-      handleContactDialogClose();
-    } catch (error) {
-      console.error('Error sending email:', error);
-      setSnackbar({
-        open: true,
-        message: 'Error sending message. Please try again later.',
-        severity: 'error',
-      });
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  const [selectedTag, setSelectedTag] = useState<'driver' | 'contactUs'>('driver');
 
   // Use the appropriate query hook based on role
   const { useAdminContacts } = useAdminQueries();
@@ -189,21 +72,20 @@ const AdminContactsPage: React.FC = () => {
     );
   }
 
-  const handleChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
-    setSelectedTag(event.target.value as 'driver' | 'contactUs');
-    if (event.target.value === 'contactUs') {
-      handleContactDialogOpen();
-      setSelectedTag('driver');
-    }
-  };
 
   return (
     <Box>
-      <NativeSelect value={selectedTag} onChange={handleChange}>
-        <option aria-label="None" value="" />
-        <option value={'driver'}>Drivers</option>
-        <option value={'contactUs'}>Contact Us</option>
-      </NativeSelect>
+      <Box
+  sx={{
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'left',
+    mb: 2,
+  }}
+>
+
+
+</Box>
       <List>
         {data?.data?.map((contact: Contact) => {
           if (contact.tag === 'driver')
@@ -278,99 +160,6 @@ const AdminContactsPage: React.FC = () => {
             );
         })}
       </List>
-      <Dialog open={openContactDialog} onClose={handleContactDialogClose} maxWidth="sm" fullWidth>
-        <DialogTitle
-          sx={{
-            backgroundColor: '#1a1a1a',
-            color: 'white',
-            fontWeight: 'bold',
-          }}
-        >
-          Contact Us
-        </DialogTitle>
-        <DialogContent sx={{ mt: 2 }}>
-          <DialogContentText sx={{ mb: 2 }}>
-            Please fill out the form below and we'll get back to you as soon as possible.
-          </DialogContentText>
-          <TextField
-            autoFocus
-            name="name"
-            label="Name"
-            type="text"
-            fullWidth
-            variant="outlined"
-            value={contactForm.name}
-            onChange={handleContactFormChange}
-            sx={{ mb: 2 }}
-          />
-          <TextField
-            name="email"
-            label="Email"
-            type="email"
-            fullWidth
-            variant="outlined"
-            value={contactForm.email}
-            onChange={handleContactFormChange}
-            sx={{ mb: 2 }}
-          />
-          <TextField
-            name="message"
-            label="Message"
-            type="text"
-            fullWidth
-            multiline
-            rows={4}
-            variant="outlined"
-            value={contactForm.message}
-            onChange={handleContactFormChange}
-          />
-        </DialogContent>
-        <DialogActions sx={{ p: 3 }}>
-          <Button
-            onClick={handleContactDialogClose}
-            variant="outlined"
-            sx={{
-              borderColor: '#fecc04',
-              color: 'black',
-              '&:hover': {
-                borderColor: '#e5b800',
-                backgroundColor: 'rgba(254, 204, 4, 0.1)',
-              },
-            }}
-          >
-            Cancel
-          </Button>
-          <Button
-            onClick={handleContactSubmit}
-            variant="contained"
-            sx={{
-              backgroundColor: '#fecc04',
-              color: 'black',
-              '&:hover': {
-                backgroundColor: '#e5b800',
-              },
-            }}
-          >
-            {isSubmitting ? 'Sending... ' : 'Send Message'}
-          </Button>
-        </DialogActions>
-      </Dialog>
-
-      <Snackbar
-        open={snackbar.open}
-        autoHideDuration={6000}
-        onClose={() => setSnackbar({ ...snackbar, open: false })}
-        anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-      >
-        <CustomAlert
-          onClose={() => setSnackbar({ ...snackbar, open: false })}
-          severity={snackbar.severity}
-          variant="filled"
-          sx={{ width: '100%' }}
-        >
-          {snackbar.message}
-        </CustomAlert>
-      </Snackbar>
     </Box>
   );
 };
