+++ src/types/feedback.types.ts
@@ -1,0 +1,5 @@
+export interface feedback{
+    name: string;
+    email: string;
+    message: string;
+  }
