+++ src/types/snackbarAlert.types.ts
@@ -1,0 +1,5 @@
+export interface SnackbarAlert {
+    open: boolean;
+    message: string;
+    severity: 'success' | 'error' | 'warning' | 'info';
+  }
